+++ src/main/java/com/commercetools/sync/categories/utils/CategoryAssetUpdateActionUtils.java
@@ -16,10 +16,9 @@ import io.sphere.sdk.models.LocalizedString;
 import javax.annotation.Nonnull;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 import static com.commercetools.sync.commons.utils.CommonTypeUpdateActionUtils.buildUpdateAction;
-import static java.util.Arrays.asList;
+import static com.commercetools.sync.commons.utils.OptionalUtils.filterEmptyOptionals;
 
 public final class CategoryAssetUpdateActionUtils {
 
@@ -39,36 +38,17 @@ public final class CategoryAssetUpdateActionUtils {
                                                             @Nonnull final AssetDraft newAsset,
                                                             @Nonnull final CategorySyncOptions syncOptions) {
 
-        final List<UpdateAction<Category>> updateActions = buildUpdateActionsFromOptionals(asList(
+        final List<UpdateAction<Category>> updateActions = filterEmptyOptionals(
             buildChangeAssetNameUpdateAction(oldAsset, newAsset),
             buildSetAssetDescriptionUpdateAction(oldAsset, newAsset),
             buildSetAssetTagsUpdateAction(oldAsset, newAsset),
             buildSetAssetSourcesUpdateAction(oldAsset, newAsset)
-        ));
+        );
 
         updateActions.addAll(buildCustomUpdateActions(oldAsset, newAsset, syncOptions));
         return updateActions;
     }
 
-    /**
-     * TODO: THIS WILL BE REMOVED AS SOON AS GITHUB ISSUE#255 is resolved.
-     * Given a list of category {@link UpdateAction} elements, where each is wrapped in an {@link Optional}; this method
-     * filters out the optionals which are only present and returns a new list of category {@link UpdateAction}
-     * elements.
-     *
-     * @param optionalUpdateActions list of category {@link UpdateAction} elements,
-     *                              where each is wrapped in an {@link Optional}.
-     * @return a List of category update actions from the optionals that were present in the
-     *         {@code optionalUpdateActions} list parameter.
-     */
-    @Nonnull
-    private static List<UpdateAction<Category>> buildUpdateActionsFromOptionals(
-        @Nonnull final List<Optional<UpdateAction<Category>>> optionalUpdateActions) {
-        return optionalUpdateActions.stream()
-                                    .filter(Optional::isPresent)
-                                    .map(Optional::get)
-                                    .collect(Collectors.toList());
-    }
 
     /**
      * Compares the {@link LocalizedString} names of an {@link Asset} and an {@link AssetDraft} and returns an